+++ lib/internal/bootstrap/node.js
@@ -18,7 +18,6 @@
 
 const { internalBinding, NativeModule } = loaderExports;
 
-const exceptionHandlerState = { captureFn: null };
 let getOptionValue;
 
 function startup() {
@@ -26,8 +25,23 @@ function startup() {
 
   setupProcessObject();
 
-  // Do this good and early, since it handles errors.
-  setupProcessFatal();
+  // TODO(joyeecheung): this does not have to done so early, any fatal errors
+  // thrown before user code execution should simply crash the process
+  // and we do not care about any clean up at that point. We don't care
+  // about emitting any events if the process crash upon bootstrap either.
+  {
+    const {
+      fatalException,
+      setUncaughtExceptionCaptureCallback,
+      hasUncaughtExceptionCaptureCallback
+    } = NativeModule.require('internal/process/execution');
+
+    process._fatalException = fatalException;
+    process.setUncaughtExceptionCaptureCallback =
+      setUncaughtExceptionCaptureCallback;
+    process.hasUncaughtExceptionCaptureCallback =
+      hasUncaughtExceptionCaptureCallback;
+  }
 
   setupGlobalVariables();
 
@@ -83,9 +97,7 @@ function startup() {
     process.reallyExit = rawMethods.reallyExit;
     process._kill = rawMethods._kill;
 
-    const wrapped = perThreadSetup.wrapProcessMethods(
-      rawMethods, exceptionHandlerState
-    );
+    const wrapped = perThreadSetup.wrapProcessMethods(rawMethods);
     process._rawDebug = wrapped._rawDebug;
     process.hrtime = wrapped.hrtime;
     process.hrtime.bigint = wrapped.hrtimeBigInt;
@@ -93,10 +105,6 @@ function startup() {
     process.memoryUsage = wrapped.memoryUsage;
     process.kill = wrapped.kill;
     process.exit = wrapped.exit;
-    process.setUncaughtExceptionCaptureCallback =
-      wrapped.setUncaughtExceptionCaptureCallback;
-    process.hasUncaughtExceptionCaptureCallback =
-      wrapped.hasUncaughtExceptionCaptureCallback;
   }
 
   NativeModule.require('internal/process/warning').setup();
@@ -305,7 +313,7 @@ function startExecution() {
   // This means we are in a Worker context, and any script execution
   // will be directed by the worker module.
   if (internalBinding('worker').getEnvMessagePort() !== undefined) {
-    NativeModule.require('internal/worker').setupChild(evalScript);
+    NativeModule.require('internal/worker').setupChild();
     return;
   }
 
@@ -376,7 +384,9 @@ function executeUserCode() {
       addBuiltinLibsToObject
     } = NativeModule.require('internal/modules/cjs/helpers');
     addBuiltinLibsToObject(global);
-    evalScript('[eval]', wrapForBreakOnFirstLine(getOptionValue('--eval')));
+    const source = getOptionValue('--eval');
+    const { evalScript } = NativeModule.require('internal/process/execution');
+    evalScript('[eval]', source, process._breakFirstLine);
     return;
   }
 
@@ -430,7 +440,8 @@ function executeUserCode() {
 
     // User passed '-e' or '--eval' along with `-i` or `--interactive`
     if (process._eval != null) {
-      evalScript('[eval]', wrapForBreakOnFirstLine(process._eval));
+      const { evalScript } = NativeModule.require('internal/process/execution');
+      evalScript('[eval]', process._eval, process._breakFirstLine);
     }
     return;
   }
@@ -452,7 +463,8 @@ function readAndExecuteStdin() {
       checkScriptSyntax(code, '[stdin]');
     } else {
       process._eval = code;
-      evalScript('[stdin]', wrapForBreakOnFirstLine(process._eval));
+      const { evalScript } = NativeModule.require('internal/process/execution');
+      evalScript('[stdin]', process._eval, process._breakFirstLine);
     }
   });
 }
@@ -654,95 +666,6 @@ function setupDOMException() {
   const DOMException = NativeModule.require('internal/domexception');
   const { registerDOMException } = internalBinding('messaging');
   registerDOMException(DOMException);
-}
-
-function noop() {}
-
-function setupProcessFatal() {
-  const {
-    executionAsyncId,
-    clearDefaultTriggerAsyncId,
-    clearAsyncIdStack,
-    hasAsyncIdStack,
-    afterHooksExist,
-    emitAfter
-  } = NativeModule.require('internal/async_hooks');
-
-  process._fatalException = (er) => {
-    // It's possible that defaultTriggerAsyncId was set for a constructor
-    // call that threw and was never cleared. So clear it now.
-    clearDefaultTriggerAsyncId();
-
-    if (exceptionHandlerState.captureFn !== null) {
-      exceptionHandlerState.captureFn(er);
-    } else if (!process.emit('uncaughtException', er)) {
-      // If someone handled it, then great.  otherwise, die in C++ land
-      // since that means that we'll exit the process, emit the 'exit' event.
-      try {
-        if (!process._exiting) {
-          process._exiting = true;
-          process.exitCode = 1;
-          process.emit('exit', 1);
-        }
-      } catch {
-        // Nothing to be done about it at this point.
-      }
-      try {
-        const { kExpandStackSymbol } = NativeModule.require('internal/util');
-        if (typeof er[kExpandStackSymbol] === 'function')
-          er[kExpandStackSymbol]();
-      } catch {
-        // Nothing to be done about it at this point.
-      }
-      return false;
-    }
-
-    // If we handled an error, then make sure any ticks get processed
-    // by ensuring that the next Immediate cycle isn't empty.
-    NativeModule.require('timers').setImmediate(noop);
-
-    // Emit the after() hooks now that the exception has been handled.
-    if (afterHooksExist()) {
-      do {
-        emitAfter(executionAsyncId());
-      } while (hasAsyncIdStack());
-    // Or completely empty the id stack.
-    } else {
-      clearAsyncIdStack();
-    }
-
-    return true;
-  };
-}
-
-function wrapForBreakOnFirstLine(source) {
-  if (!process._breakFirstLine)
-    return source;
-  const fn = `function() {\n\n${source};\n\n}`;
-  return `process.binding('inspector').callAndPauseOnStart(${fn}, {})`;
-}
-
-function evalScript(name, body) {
-  const CJSModule = NativeModule.require('internal/modules/cjs/loader');
-  const path = NativeModule.require('path');
-  const { tryGetCwd } = NativeModule.require('internal/util');
-  const cwd = tryGetCwd(path);
-
-  const module = new CJSModule(name);
-  module.filename = path.join(cwd, name);
-  module.paths = CJSModule._nodeModulePaths(cwd);
-  const script = `global.__filename = ${JSON.stringify(name)};\n` +
-                  'global.exports = exports;\n' +
-                  'global.module = module;\n' +
-                  'global.__dirname = __dirname;\n' +
-                  'global.require = require;\n' +
-                  'return require("vm").runInThisContext(' +
-                  `${JSON.stringify(body)}, { filename: ` +
-                  `${JSON.stringify(name)}, displayErrors: true });\n`;
-  const result = module._compile(script, `${name}-wrapper`);
-  if (getOptionValue('--print')) console.log(result);
-  // Handle any nextTicks added in the first tick of the program.
-  process._tickCallback();
 }
 
 function checkScriptSyntax(source, filename) {
