+++ lib/internal/console/inspector.js
@@ -1,15 +1,14 @@
 'use strict';
 
-const path = require('path');
 const CJSModule = require('internal/modules/cjs/loader');
 const { makeRequireFunction } = require('internal/modules/cjs/helpers');
-const { tryGetCwd } = require('internal/util');
+const { tryGetCwd } = require('internal/process/execution');
 const { addCommandLineAPI, consoleCall } = internalBinding('inspector');
 
 // Wrap a console implemented by Node.js with features from the VM inspector
 function addInspectorApis(consoleFromNode, consoleFromVM) {
   // Setup inspector command line API.
-  const cwd = tryGetCwd(path);
+  const cwd = tryGetCwd();
   const consoleAPIModule = new CJSModule('<inspector console>');
   consoleAPIModule.paths =
       CJSModule._nodeModulePaths(cwd).concat(CJSModule.globalPaths);
