+++ lib/internal/process/per_thread.js
@@ -12,7 +12,6 @@ const {
     ERR_INVALID_ARG_TYPE,
     ERR_INVALID_OPT_VALUE,
     ERR_OUT_OF_RANGE,
-    ERR_UNCAUGHT_EXCEPTION_CAPTURE_ALREADY_SET,
     ERR_UNKNOWN_SIGNAL
   }
 } = require('internal/errors');
@@ -24,7 +23,7 @@ function assert(x, msg) {
 }
 
 // The execution of this function itself should not cause any side effects.
-function wrapProcessMethods(binding, exceptionHandlerState) {
+function wrapProcessMethods(binding) {
   const {
     hrtime: _hrtime,
     hrtimeBigInt: _hrtimeBigInt,
@@ -185,29 +184,6 @@ function wrapProcessMethods(binding, exceptionHandlerState) {
     return true;
   }
 
-  // shouldAbortOnUncaughtToggle is a typed array for faster
-  // communication with JS.
-  const { shouldAbortOnUncaughtToggle } = binding;
-
-  function setUncaughtExceptionCaptureCallback(fn) {
-    if (fn === null) {
-      exceptionHandlerState.captureFn = fn;
-      shouldAbortOnUncaughtToggle[0] = 1;
-      return;
-    }
-    if (typeof fn !== 'function') {
-      throw new ERR_INVALID_ARG_TYPE('fn', ['Function', 'null'], fn);
-    }
-    if (exceptionHandlerState.captureFn !== null) {
-      throw new ERR_UNCAUGHT_EXCEPTION_CAPTURE_ALREADY_SET();
-    }
-    exceptionHandlerState.captureFn = fn;
-    shouldAbortOnUncaughtToggle[0] = 0;
-  }
-
-  function hasUncaughtExceptionCaptureCallback() {
-    return exceptionHandlerState.captureFn !== null;
-  }
 
   return {
     _rawDebug,
@@ -216,9 +192,7 @@ function wrapProcessMethods(binding, exceptionHandlerState) {
     cpuUsage,
     memoryUsage,
     kill,
-    exit,
-    setUncaughtExceptionCaptureCallback,
-    hasUncaughtExceptionCaptureCallback
+    exit
   };
 }
