+++ lib/internal/util.js
@@ -387,17 +387,6 @@ function once(callback) {
   };
 }
 
-function tryGetCwd(path) {
-  try {
-    return process.cwd();
-  } catch {
-    // getcwd(3) can fail if the current working directory has been deleted.
-    // Fall back to the directory name of the (absolute) executable path.
-    // It's not really correct but what are the alternatives?
-    return path.dirname(process.execPath);
-  }
-}
-
 module.exports = {
   assertCrypto,
   cachedResult,
@@ -417,7 +406,6 @@ module.exports = {
   once,
   promisify,
   spliceOne,
-  tryGetCwd,
   removeColors,
 
   // Symbol used to customize promisify conversion
