+++ lib/internal/worker.js
@@ -432,7 +432,7 @@ if (!isMainThread) {
 
 let originalFatalException;
 
-function setupChild(evalScript) {
+function setupChild() {
   // Called during bootstrap to set up worker script execution.
   debug(`[${threadId}] is setting up worker child environment`);
   const port = getEnvMessagePort();
@@ -453,6 +453,7 @@ function setupChild(evalScript) {
       port.unref();
       port.postMessage({ type: messageTypes.UP_AND_RUNNING });
       if (doEval) {
+        const { evalScript } = require('internal/process/execution');
         evalScript('[worker eval]', filename);
       } else {
         process.argv[1] = filename; // script filename
